Add protection from server errors
Sometimes application raise IncompleteRead error (and some others as a result). Since it is unpredictable server-side error, we just reload the parsing of the same site from the beginning. Runtime tests show that errors do not repeat more than three times for the same site, so repeat parsing until success is quite unharmed for the process.